src/main/java/edu/ifma/locacaodeimoveis/repository/ClienteRepository.java
package edu.ifma.locacaodeimoveis.repository;

import edu.ifma.locacaodeimoveis.model.Cliente;

import javax.persistence.EntityManager;

public class ClienteRepository {

    private final EntityManager manager;
    private final DAOGenerico<Cliente> daoGenerico;

    public ClienteRepository(EntityManager manager) {
        this.manager = manager;
        this.daoGenerico = new DAOGenerico<>(manager);
    }

    public Cliente buscaPor(String nome) {
        return manager.createQuery("From Cliente c where c.nomeCliente = :nome", Cliente.class)
                .setParameter("nome", nome)
                .getSingleResult();
    }

    public Cliente salvaOuAtualiza(Cliente cliente) {
        return daoGenerico.salvaOuAtualiza(cliente);
    }

    public void remove(Cliente cliente) {
        daoGenerico.remove(cliente);
    }

}
